[BUG](mdac): Fix Modal deployment
Avoid shadowing the Modal package with the deployment module name.\nUpdate the deployment definition to the current Modal API and document\nthe staging command and remote Cargo image build.
